New toolchain support in make rules (#117)
* Enable toolchain detection, prefer mipsel-unknown-elf. 
* Launch script should also recognize available custom toolchain
* Fixed address type sizes
* Using %zu for printing out size_t

+++ sys/callout.c
@@ -41,12 +41,12 @@ void callout_setup(callout_t *handle, realtime_t time, timeout_t fn,
   handle->c_index = index;
   callout_set_pending(handle);
 
-  log("Add callout {%p} with wakeup at %ld.", handle, (size_t)handle->c_time);
+  log("Add callout {%p} with wakeup at %lld.", handle, handle->c_time);
   TAILQ_INSERT_TAIL(&ci.heads[index], handle, c_link);
 }
 
 void callout_stop(callout_t *handle) {
-  log("Remove callout {%p} at %ld.", handle, (size_t)handle->c_time);
+  log("Remove callout {%p} at %lld.", handle, handle->c_time);
   TAILQ_REMOVE(&ci.heads[handle->c_index], handle, c_link);
 }
